delta_sysfs.c: fix a compiler warning
A warning emitted by gcc-4.1.2 on x86:

delta_sysfs.c: In function ‘ploop_find_dev_by_delta’:
delta_sysfs.c:437: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 4 has type ‘dev_t’
delta_sysfs.c:437: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 5 has type ‘__dev_t’

+++ lib/delta_sysfs.c
@@ -434,7 +434,7 @@ int ploop_find_dev_by_delta(char *delta, char *buf, int size)
 					st1.st_rdev != dev)
 			{
 				ploop_err(0, "Inconsistency in device number detected for %s sys_dev=%lu dev=%lu",
-						buf, dev, st1.st_rdev);
+						buf, (unsigned long)dev, (unsigned long)st1.st_rdev);
 				ret = -1;
 			} else
 				ret = 0;
